+++ test/OpenTelemetry.Exporter.Prometheus.HttpListener.Tests/PrometheusEscapingTests.cs
@@ -13,8 +13,10 @@ public static class PrometheusEscapingTests
     [InlineData("mysystem.prod.west.cpu.load_total", "mysystem_dot_prod_dot_west_dot_cpu_dot_load__total")]
     [InlineData("http.status:sum", "http_dot_status:sum")]
     [InlineData("no:escaping_required", "no:escaping__required")]
-    [InlineData("label with \U0001F631", "label__with____")]
-    [InlineData("\u82B1\u706B", "____")]
+    [InlineData("a_b", "a__b")]
+    [InlineData("a b", "a_b")]
+    [InlineData("label with \U0001F631", "label_with__")]
+    [InlineData("\u82B1\u706B", "__")]
     [InlineData("", "")]
     public static void EscapeName_Dots(string name, string expected)
         => Assert.Equal(expected, PrometheusEscaping.EscapeName(name, EscapingScheme.Dots));
@@ -25,6 +27,8 @@ public static class PrometheusEscapingTests
     [InlineData("mysystem.prod.west.cpu.load_total", "U__mysystem_2e_prod_2e_west_2e_cpu_2e_load__total")]
     [InlineData("http.status:sum", "U__http_2e_status:sum")]
     [InlineData("no:escaping_required", "no:escaping_required")]
+    [InlineData("foo.bar", "U__foo_2e_bar")]
+    [InlineData("U__foo_2e_bar", "U___55_____foo__2e__bar")]
     [InlineData("label with \u0100", "U__label_20_with_20__100_")]
     [InlineData("", "")]
     public static void EscapeName_Values(string name, string expected)
@@ -102,8 +106,8 @@ public static class PrometheusEscapingTests
         => Assert.Equal(expected, PrometheusEscaping.IsValidLegacyLabelName(name));
 
     [Theory]
-    [InlineData("http:method", "http__method")]
-    [InlineData("http.method:value", "http_dot_method__value")]
+    [InlineData("http:method", "http_method")]
+    [InlineData("http.method:value", "http_dot_method_value")]
     public static void EscapeName_Dots_LabelNameEncodesColon(string name, string expected)
         => Assert.Equal(expected, PrometheusEscaping.EscapeName(name, EscapingScheme.Dots, isMetricName: false));