+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -30,6 +30,10 @@ Notes](../../RELEASENOTES.md).
 * Fixed the OTLP/gRPC exporter logging incorrectly when an export timed out.
   ([#7455](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7455))
 
+* The `Resource` Schema URL is now exported on the OTLP `ResourceSpans`,
+  `ResourceMetrics`, and `ResourceLogs` messages.
+  ([#7472](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7472))
+
 ## 1.16.0
 
 Released 2026-Jun-10